requirements.txt
astroid==2.0.4
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
nano==0.9.4
pylint==2.1.1
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
wrapt==1.10.11
